feat(general/repo): add material database

+++ services/general/migration/000002_create_material.up.sql
@@ -9,7 +9,7 @@ CREATE TYPE general.material_kind AS ENUM (
 CREATE TABLE IF NOT EXISTS general.materials(
     id uuid PRIMARY KEY DEFAULT uuid_generate_v4(),
     course_id uuid NOT NULL REFERENCES general.courses(id) ON DELETE CASCADE,
-    position numeric NOT NULL,
+    position double precision NOT NULL,
     kind general.material_kind NOT NULL,
     lesson text NOT NULL,
     title text NOT NULL,
